Partial update to documentation
Updated three pages in Advanced Topics section that is currently not
being published. The Introduction and Custom Plug-ins pages are ready
for publication after technical validation. The Quantum HA page has no
content and the final page contains a note that indicates it needs to be
edited, though I think that it should be validated for technical
accuracy first.