Bulk edit and updates
This is about half of the documentation re-checked and verified. Made
numerous small changes and fixes as well as modified a few incorrect
references to support the changes. Added missing RST extension to the
add remove nodes FAQ entry and renamed 0030-mysql-pacemaker.rst to
0040-corosync-crashes.rst as the first name was completely incorrect and
the file number was conflicting with an existing file.