General arch, added Ceph and Galera info
A quick fix to reflect that we no longer have separate multi-node and
multi-node HA models in 5.0.

Also moved information about Ceph and MySQL/Galera from FAQ doc
and augmented it with information from Miroslav Anashkin and
Dmitriy Novakovskiy.

This removes the last bits of the FAQ, which is now completely removed
from the landing page.

The "bigpicture" architectural description needs additional work
but this package is reasonable accurate as far as it goes.

Co-Authored-By: Dmitry Borodaenko <dborodaenko@mirantis.com>
Change-Id: Ia29d906a8d1377c08c822262f4d3759b9fd84593