Replace intro material for Planning and User Guides
This replaces 94510

Changes for new intro:
- pre-install -- 0010-intro text is replaced with new material
- user-guide -- 0000-intro text is replaced with new material

- install-guide -- deleted 0000-intro-header and the contents of
  the 0000-intro subdirectory (they were deleted in 94510 because
  they are no longer needed)

This also pulls install-guide out of the build.  This is necessary
to make the 5.0 release and is not a permanent change so unused
files such as those in the contents/ and pdf/ directories have
not been removed:
- contents.rst
- index.rst
- install-guide.rst
- pdf/conf.py -- had to delete the install-guide line

Modified some top-level *.rst files trying to resolve build and
xref issues -- some of these tags had _ rather than - which causes
problems.

Change-Id: I9ab382034caf2ecb7aaa9c552d10552e97be89cf