Ceilometer and MongoDB configuration information
planning/monitoring/ceilometer-mongodb section, which is based on
the old Ceilometer chapter that was in the Ops Guide.
Add information about determining resoures for Ceilometer and MongoDB.

operations/monitoring/8500-ceilometer section to discuss
the default configuration that Fuel deploys
and how to configure and use Ceilometer.
This section hands off to community docs for details.

terminology/ceilometer -- fixed xrefs

Closes-Bug: 1357144

Change-Id: I8de9d58058018a187918e54827d5df826f2c6e4f