Hardware requirements for target nodes
planning-guide
- incorporated introductory material into 0020-system-requirements;
  deleted 0021-sysreq-intro file

- added general hardware guidelines to 0020-system-requirements

- 0020-system-requirements/0023-node-server-hardware-requirements
  - added links to community docs for Compute and Image storage requirements,
    object storage requirements
  - added link to OpenStack Operations Guide
  - added link to "nodes and roles" and monitoring/ceilometer
  - added link to "Calculating Hardware" section of Planning Guide

- added "Sample Target Node Configuration" based on info from Stefan

The two files in 0010-prerequisites suddently had Title-level build
errors so I fixed those.

user-guide
  Changed the recommendation for Ceph replication factor to 3, to
  match the recommendation given here.

terminology/overcommit -- added note to set CPU ratio to 1:1
until one understands the resource allocation for ones environment,
in support of note in target nodes recommendations.
We need a larger discussion about how to determine an appropriate
overcommit ratio but this will be more appropriate when one can reset
the ratio on the UI and without redeploying the environment.  It is
beyond the scope of this commit.

Closes-Bug: 1373419

Change-Id: I5f6e89646fccbaf1e77b1ef14cc154e4d78d290d