Add statistics screen to User Guide
Adds a section about statistics gathering
to the "Create Environment" section of the user guide.

Note that the Enterprise and Community editions of MOS
have different versions of this screen.  Patch 2 includes
the Community version.  Patches 1 and 3 contain the
Enterprise version; Patch 3 adds a note that the Community
version has a different screen.

Change-Id: I36835b1f430a9d87d1584fedacb2e79d14a7ce11