Move product description to index.rst from Release Notes
This moves the "What is Mirantis OpenStack" information from
the Release Notes to the index.rst file; it plays out as three
items in the main panel of the docs landing page,
all of which are links to articles in the Terminology Reference
with a brief bit of supporting text.

The information is similar to what was in the Release Notes
but I made a few modifications to merge info with existing
info in the Terminology Guide.

The terminology/fuel.rst article now has a paragraph about
plug-ins.  This needs links to the developer and user docs
after that is merged.

Change-Id: Ib3721d72f2630d139448405ecd0b2977b371a2ba