Modularize new features in Release Notes
STRUCTURE MODIFICATIONS ONLY! No text changes have been made.

Breaking up the New Features and Other Enhancements sections
of the Release Notes so we have one file per feature heading.
This will enable us to populate this information from the CRs
about the feature so the Release Notes can cross-reference the
other docs.

"Other Enhancements" was not in the tree from the TP so added that
and created the other-enhancements subdirectory.  Content will be
developed in the same modular fashion as the new-features.

Change-Id: I8b00c1a2c946de17133b9abed71e193679d89159