How to add MongoDB nodes using shell commands
This should be merged to master and stable/6.0

Added "Add MongoDB nodes" to Ops-Guide section about redeploying nodes

Added link to that section from Release Notes; reworded Release Notes to
say MongoDB nodes can not be added using the Fuel Web UI.

Added information to the fuel-ref/astute.yaml page about
the ceilometer/mongodb settings that are referenced from this procedure.

I added the full contents of the Target astute.yaml file and did a quick
pass to divide it into sections that can be discussed reasonably; most of
the work on this reference page will be done in a separate CR.

I did fix the file-ref doc so the TOC only shows the top-level section
(the file name) rather than all the subsections.

Change-Id: I237f15f7f57b2b4532566ea74fc6ca701124fee8
Partial-Bug: 1308990