Remove support for external MongoDB server from Release Notes
This feature was moved to 6.1 but is included in the "New Features"
list of the Release Notes.  I left the file in place, just commented
out the line that sources it, since we expect to include this feature
in a later release.

Change-Id: Iae8c444927a63be76f6dea116c5fd776c32657d4
(cherry picked from commit 387b96711152e6e208914ba3cd1bbaced39378cc)