Structure 6.1 Release Notes
Provides a basic structure for the 6.1 Release Notes.

Content will be added/modified in separate CRs

Includes headlines for major features as understood at the this time
with links to specs for the feature; the wording of these headlines
will probably be modified and the xrefs to the specs will mostly be
replaced by links to the documentation that describes the feature.

Headlines for partner features are not included in this CR.

Change-Id: I76953d6f4bd0032b7ca8b9776ef391684b730bac